+++ CHANGELOG.md
@@ -18,6 +18,9 @@ The format will be:
 
 ## [Unreleased]
 
+### Changed
+- Logs detailed traceback information in case of an error to the database.
+
 ### Fixed
 - Updated the favicon's location to an absolute file path.
 - Fixed an error where requests weren't being logged in the case of unsuccessful requests.